feat: add -t, --push-tags flag to deploy
If flag is provided, gamma will verify no tag exists and tag the commit based on the package.version field of each action.

Additionally, a new `check-versions` command is added to verify version does not exist yet for changed versions in a PR (should be squashed commit to verify the PR).